refactor: tidy all imports and make sure file names are always used
@affects atoms, core, immer, react
