+++ packages/atoms/src/injectors/injectMemo.ts
@@ -1,6 +1,6 @@
 import { createInjector } from '../factories/createInjector'
-import { InjectorDeps, PartialAtomInstance } from '../types'
-import { haveDepsChanged, prefix } from '../utils'
+import { InjectorDeps, PartialAtomInstance } from '../types/index'
+import { haveDepsChanged, prefix } from '../utils/index'
 
 type MemoInjectorDescriptor<T> = {
   deps: InjectorDeps
