+++ packages/atoms/src/injectors/injectStore.ts
@@ -1,7 +1,7 @@
 import { createStore, internalTypes, Store } from '@zedux/core'
 import { createInjector } from '../factories/createInjector'
-import { InjectStoreConfig, PartialAtomInstance } from '../types'
-import { InjectorDescriptor, prefix } from '../utils'
+import { InjectStoreConfig, PartialAtomInstance } from '../types/index'
+import { InjectorDescriptor, prefix } from '../utils/index'
 
 export const doSubscribe = <State>(
   instance: PartialAtomInstance,
