+++ packages/atoms/src/types/atoms.ts
@@ -5,16 +5,14 @@ import { AtomInstanceBase } from '../classes/instances/AtomInstanceBase'
 
 type AnyAtomGenerics = { [K in keyof AtomGenerics]: any }
 
-export type AnyAtomInstance<
-  G extends Partial<AtomGenerics> | 'any' = 'any'
-> = AtomInstance<G extends Partial<AtomGenerics> ? AtomGenericsPartial<G> : any>
+export type AnyAtomInstance<G extends Partial<AtomGenerics> | 'any' = 'any'> =
+  AtomInstance<G extends Partial<AtomGenerics> ? AtomGenericsPartial<G> : any>
 
-export type AnyAtomTemplate<
-  G extends Partial<AtomGenerics> | 'any' = 'any'
-> = AtomTemplateBase<
-  G extends Partial<AtomGenerics> ? AtomGenericsPartial<G> : any,
-  AnyAtomInstance<G>
->
+export type AnyAtomTemplate<G extends Partial<AtomGenerics> | 'any' = 'any'> =
+  AtomTemplateBase<
+    G extends Partial<AtomGenerics> ? AtomGenericsPartial<G> : any,
+    AnyAtomInstance<G>
+  >
 
 export interface AtomGenerics {
   Exports: Record<string, any>
@@ -32,49 +30,43 @@ export type AtomGenericsPartial<G extends Partial<AtomGenerics>> = Omit<
 
 export type AtomApiPromise = Promise<any> | undefined
 
-export type AtomExportsType<
-  A extends AnyAtomTemplate | AnyAtomInstance
-> = A extends AtomTemplateBase<infer G, any>
-  ? G['Exports']
-  : A extends AtomInstance<infer G>
-  ? G['Exports']
-  : never
+export type AtomExportsType<A extends AnyAtomTemplate | AnyAtomInstance> =
+  A extends AtomTemplateBase<infer G, any>
+    ? G['Exports']
+    : A extends AtomInstance<infer G>
+    ? G['Exports']
+    : never
 
-export type AtomInstanceType<
-  A extends AnyAtomTemplate
-> = A extends AtomTemplateBase<any, infer T> ? T : never
+export type AtomInstanceType<A extends AnyAtomTemplate> =
+  A extends AtomTemplateBase<any, infer T> ? T : never
 
-export type AtomParamsType<
-  A extends AnyAtomTemplate | AnyAtomInstance
-> = A extends AtomTemplateBase<infer G, AtomInstance<infer G>>
-  ? G['Params']
-  : A extends AtomInstance<infer G>
-  ? G['Params']
-  : never
+export type AtomParamsType<A extends AnyAtomTemplate | AnyAtomInstance> =
+  A extends AtomTemplateBase<infer G, AtomInstance<infer G>>
+    ? G['Params']
+    : A extends AtomInstance<infer G>
+    ? G['Params']
+    : never
 
-export type AtomPromiseType<
-  A extends AnyAtomTemplate | AnyAtomInstance
-> = A extends AtomTemplateBase<infer G, AtomInstance<infer G>>
-  ? G['Promise']
-  : A extends AtomInstance<infer G>
-  ? G['Promise']
-  : never
+export type AtomPromiseType<A extends AnyAtomTemplate | AnyAtomInstance> =
+  A extends AtomTemplateBase<infer G, AtomInstance<infer G>>
+    ? G['Promise']
+    : A extends AtomInstance<infer G>
+    ? G['Promise']
+    : never
 
-export type AtomStateType<
-  A extends AnyAtomTemplate | AnyAtomInstance
-> = A extends AtomTemplateBase<infer G, AtomInstance<infer G>>
-  ? G['State']
-  : A extends AtomInstance<infer G>
-  ? G['State']
-  : never
+export type AtomStateType<A extends AnyAtomTemplate | AnyAtomInstance> =
+  A extends AtomTemplateBase<infer G, AtomInstance<infer G>>
+    ? G['State']
+    : A extends AtomInstance<infer G>
+    ? G['State']
+    : never
 
-export type AtomStoreType<
-  A extends AnyAtomTemplate | AnyAtomInstance
-> = A extends AtomTemplateBase<infer G, AtomInstance<infer G>>
-  ? G['Store']
-  : A extends AtomInstance<infer G>
-  ? G['Store']
-  : never
+export type AtomStoreType<A extends AnyAtomTemplate | AnyAtomInstance> =
+  A extends AtomTemplateBase<infer G, AtomInstance<infer G>>
+    ? G['Store']
+    : A extends AtomInstance<infer G>
+    ? G['Store']
+    : never
 
 export type AtomTemplateType<
   A extends AtomInstanceBase<any, AtomTemplateBase<any, AtomInstance<any>>>
