+++ packages/react/src/index.ts
@@ -1,4 +1,4 @@
 export * from '@zedux/atoms'
-export * from './components'
-export * from './hooks'
+export * from './components/index'
+export * from './hooks/index'
 export * from './types'
